Add Frankfurter API as default Forex Pricing Provider
